% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{.create_long_format}
\alias{.create_long_format}
\title{Create long format from wide QCI data}
\usage{
.create_long_format(wide_data)
}
\arguments{
\item{wide_data}{Wide format data.table.}
}
\value{
Long format data.table.
}
\description{
Create long format from wide QCI data
}
\keyword{internal}
