% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{location_type}
\alias{location_type}
\title{Location type metadata}
\format{
A data.frame with columns:
\describe{
\item{location_id}{Integer. GBD location identifier.}
\item{location_name}{Character. GBD location name.}
\item{type}{Character. Location classification (e.g., "Country",
"WHO Regions", "SDI", "World Bank Income Levels").}
\item{iso3}{Character. ISO 3166-1 alpha-3 country code (NA for
non-country locations).}
}
}
\usage{
location_type
}
\description{
Mapping of GBD location identifiers to location type (Country, WHO Region,
World Bank Income Level, SDI, etc.) and ISO3 codes. Used to filter
data to country-level observations and to join ISO3 codes for mapping.
}
\keyword{datasets}
