% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdr-map.R
\name{plot_gdr_map}
\alias{plot_gdr_map}
\title{Create a choropleth map of Gender Disparity Ratio}
\usage{
plot_gdr_map(
  data,
  years = NULL,
  breaks = c(0, 0.5, 0.95, 1.05, 2, Inf),
  labels = c("< 0.5", "[0.5, 0.95)", "[0.95, 1.05]", "(1.05, 2.0]", "> 2.0"),
  colors = c("#878787", "#b7b7b7", "#64c2a5", "#b1abd0", "#7f74ac"),
  title = NULL,
  legend_title = "Age-standardized GDR"
)
}
\arguments{
\item{data}{A data.frame with at least: \code{iso3}, \code{gdr}, and optionally
\code{year}. Use \code{\link[=merge_location_type]{merge_location_type()}} to add \code{iso3} to \code{\link[=qci_gdr]{qci_gdr()}}
output.}

\item{years}{Integer vector. Default \code{NULL} (all).}

\item{breaks}{Numeric vector of GDR breakpoints.
Default \code{c(0, 0.5, 0.95, 1.05, 2.0, Inf)}.}

\item{labels}{Character vector of labels for GDR bins.
Default \code{c("< 0.5", "[0.5, 0.95)", "[0.95, 1.05]", "(1.05, 2.0]", "> 2.0")}.}

\item{colors}{Character vector of 5 colors.
Default grey-to-purple scale.}

\item{title}{Character. Default \code{NULL} (auto-generated).}

\item{legend_title}{Character. Default \code{"Age-standardized GDR"}.}
}
\value{
A ggplot2 object.
}
\description{
Create a choropleth map of Gender Disparity Ratio
}
\examples{
\dontrun{
result <- qci_pipeline(sample_gbd)
gdr <- qci_gdr(result$wide)
gdr_map <- merge_location_type(gdr)
plot_gdr_map(gdr_map, years = c(1990, 2019))
}
}
