% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_measure_names}
\alias{standardize_measure_names}
\title{Standardize GBD measure names}
\usage{
standardize_measure_names(x)
}
\arguments{
\item{x}{Character vector of measure_name values.}
}
\value{
Character vector with shortened names.
}
\description{
Converts verbose GBD measure names to short forms.
}
\keyword{internal}
